refactor(scripts): share R2 transfer logic
