+++ scripts/lib/r2-orphan-audio-cleanup.mts
@@ -1,10 +1,19 @@
-import { createHash, randomUUID } from 'node:crypto';
-import { constants, createReadStream, createWriteStream } from 'node:fs';
-import { copyFile, link, mkdir, rm, stat } from 'node:fs/promises';
-import path from 'node:path';
-import { Readable, Transform } from 'node:stream';
-import { pipeline } from 'node:stream/promises';
 import { parseArgs } from 'node:util';
+
+import {
+  compareR2ObjectsOldestFirst as compareOldestFirst,
+  normalizeEtag,
+  parseByteSize,
+  type R2Client,
+  type R2DeleteResponse,
+  type R2HeadObject,
+  type R2Location,
+  type R2ObjectMetadata,
+} from './r2-transfer.mts';
+
+export type CleanupR2Client = R2Client;
+export type AllowedLocation = R2Location;
+export type { R2ObjectMetadata } from './r2-transfer.mts';
 
 export const MANIFEST_SCHEMA_VERSION = 1;
 export const MINIMUM_AGE_DAYS = 45;
@@ -12,26 +21,10 @@ export const DEFAULT_PAGE_SIZE = 1000;
 export const DEFAULT_DELETE_BATCH_SIZE = 100;
 
 const MILLISECONDS_PER_DAY = 24 * 60 * 60 * 1000;
-const SIMPLE_ETAG_PATTERN = /^[a-f\d]{32}$/i;
-const SIZE_PATTERN = /^(\d+(?:\.\d+)?)\s*(KB|MB|GB|TB|KiB|MiB|GiB|TiB)$/i;
 
 export interface CleanupConfig {
   apiBucket: string;
   mainBucket: string;
-}
-
-export interface AllowedLocation {
-  bucket: string;
-  prefix: string;
-}
-
-export interface R2ObjectMetadata {
-  bucket: string;
-  etag: string;
-  key: string;
-  lastModified: string;
-  prefix: string;
-  size: number;
 }
 
 export interface ManifestSummaryEntry {
@@ -94,50 +87,6 @@ export interface StorageKeyPageSource {
   listStorageKeys: (from: number, to: number) => Promise<string[]>;
 }
 
-export interface R2ListedObject {
-  etag?: string;
-  key?: string;
-  lastModified?: Date;
-  size?: number;
-}
-
-export interface R2ListPage {
-  nextContinuationToken?: string;
-  objects: R2ListedObject[];
-}
-
-export interface R2HeadObject {
-  etag?: string;
-  lastModified?: Date;
-  size?: number;
-}
-
-export interface R2DeleteError {
-  code?: string;
-  key: string;
-  message?: string;
-}
-
-export interface R2DeleteResponse {
-  deleted: string[];
-  errors: R2DeleteError[];
-}
-
-export interface CleanupR2Client {
-  deleteObjects: (bucket: string, keys: string[]) => Promise<R2DeleteResponse>;
-  getObject: (
-    bucket: string,
-    key: string,
-    expectedEtag: string,
-  ) => Promise<AsyncIterable<Uint8Array>>;
-  headObject: (bucket: string, key: string) => Promise<R2HeadObject | null>;
-  listObjects: (
-    bucket: string,
-    prefix: string,
-    continuationToken?: string,
-  ) => Promise<R2ListPage>;
-}
-
 export type RecheckStatus =
   | 'changed-in-r2'
   | 'eligible'
@@ -149,29 +98,6 @@ export interface RecheckResult {
   candidate: R2ObjectMetadata;
   reason?: string;
   status: RecheckStatus;
-}
-
-export interface DownloadSelection {
-  deferred: R2ObjectMetadata[];
-  selected: R2ObjectMetadata[];
-  transferBytes: number;
-}
-
-export type LocalVerificationStatus =
-  | 'missing'
-  | 'mismatch'
-  | 'unverifiable'
-  | 'verified';
-
-export interface LocalVerification {
-  reason?: string;
-  status: LocalVerificationStatus;
-}
-
-export interface DownloadResult {
-  destination: string;
-  reason?: string;
-  status: 'downloaded' | 'exists' | 'failed' | 'unverifiable';
 }
 
 export interface DeleteResult {
@@ -197,39 +123,6 @@ export function buildAllowedLocations({
   }
 
   return [...unique.values()];
-}
-
-export function parseByteSize(input: string): number {
-  const match = SIZE_PATTERN.exec(input.trim());
-  if (!match) {
-    throw new Error(
-      'Size must use KB, MB, GB, TB, KiB, MiB, GiB, or TiB, for example 20GB',
-    );
-  }
-
-  const value = Number(match[1]);
-  if (!(value > 0)) {
-    throw new Error('Size must be greater than zero');
-  }
-
-  const unit = match[2].toLowerCase();
-  const powers: Record<string, number> = {
-    gb: 1000 ** 3,
-    gib: 1024 ** 3,
-    kb: 1000,
-    kib: 1024,
-    mb: 1000 ** 2,
-    mib: 1024 ** 2,
-    tb: 1000 ** 4,
-    tib: 1024 ** 4,
-  };
-  const bytes = value * powers[unit];
-
-  if (!Number.isSafeInteger(bytes)) {
-    throw new Error('Size must resolve to a whole, safe number of bytes');
-  }
-
-  return bytes;
 }
 
 export function parseCleanupCliArgs(args: string[]): CleanupCliOptions {
@@ -319,19 +212,6 @@ export function isAtLeastDaysOld(
   return timestamp <= now.getTime() - days * MILLISECONDS_PER_DAY;
 }
 
-export function normalizeEtag(etag: string): string {
-  const trimmed = etag.trim();
-  if (trimmed.startsWith('"') && trimmed.endsWith('"')) {
-    return trimmed.slice(1, -1);
-  }
-  return trimmed;
-}
-
-export function simpleEtagMd5(etag: string): string | null {
-  const normalized = normalizeEtag(etag);
-  return SIMPLE_ETAG_PATTERN.test(normalized) ? normalized.toLowerCase() : null;
-}
-
 export function findAllowedLocation(
   bucket: string,
   key: string,
@@ -398,30 +278,6 @@ export function findOrphanCandidates(
         isAtLeastDaysOld(object.lastModified, now),
     )
     .sort(compareOldestFirst);
-}
-
-export function selectByDownloadLimit(
-  candidates: R2ObjectMetadata[],
-  maxBytes: number,
-): DownloadSelection {
-  if (!Number.isSafeInteger(maxBytes) || maxBytes <= 0) {
-    throw new Error('Download limit must be a positive, safe integer');
-  }
-
-  const selected: R2ObjectMetadata[] = [];
-  const deferred: R2ObjectMetadata[] = [];
-  let transferBytes = 0;
-
-  for (const candidate of [...candidates].sort(compareOldestFirst)) {
-    if (candidate.size <= maxBytes - transferBytes) {
-      selected.push(candidate);
-      transferBytes += candidate.size;
-    } else {
-      deferred.push(candidate);
-    }
-  }
-
-  return { deferred, selected, transferBytes };
 }
 
 export function summarizeObjects(
@@ -634,62 +490,6 @@ export async function fetchAllStorageKeys(
   return keys;
 }
 
-export async function listAllObjects(
-  client: CleanupR2Client,
-  location: AllowedLocation,
-): Promise<R2ObjectMetadata[]> {
-  const objects: R2ObjectMetadata[] = [];
-  const seenTokens = new Set<string>();
-  let continuationToken: string | undefined;
-
-  while (true) {
-    const page = await client.listObjects(
-      location.bucket,
-      location.prefix,
-      continuationToken,
-    );
-
-    for (const object of page.objects) {
-      if (
-        !object.key ||
-        object.size === undefined ||
-        !object.lastModified ||
-        !object.etag
-      ) {
-        throw new Error(
-          `R2 returned incomplete metadata under ${location.bucket}/${location.prefix}`,
-        );
-      }
-      if (!object.key.startsWith(location.prefix)) {
-        throw new Error(
-          `R2 returned an object outside requested prefix ${location.prefix}`,
-        );
-      }
-
-      objects.push({
-        bucket: location.bucket,
-        etag: normalizeEtag(object.etag),
-        key: object.key,
-        lastModified: object.lastModified.toISOString(),
-        prefix: location.prefix,
-        size: object.size,
-      });
-    }
-
-    const nextToken = page.nextContinuationToken;
-    if (!nextToken) {
-      break;
-    }
-    if (seenTokens.has(nextToken)) {
-      throw new Error('R2 pagination returned a repeated continuation token');
-    }
-    seenTokens.add(nextToken);
-    continuationToken = nextToken;
-  }
-
-  return objects;
-}
-
 export async function recheckCandidates(
   candidates: R2ObjectMetadata[],
   databaseKeys: ReadonlySet<string>,
@@ -742,152 +542,6 @@ export async function recheckCandidates(
   }
 
   return results;
-}
-
-export function resolveLocalObjectPath(
-  downloadDir: string,
-  bucket: string,
-  key: string,
-): string {
-  assertBucketName(bucket);
-  if (
-    !key ||
-    path.isAbsolute(key) ||
-    key.split(/[\\/]/).some((segment) => segment === '..')
-  ) {
-    throw new Error(`Unsafe R2 key: ${key}`);
-  }
-
-  const bucketDirectory = path.resolve(downloadDir, bucket);
-  const destination = path.resolve(bucketDirectory, key);
-  const relative = path.relative(bucketDirectory, destination);
-
-  if (!relative || relative === '..' || relative.startsWith(`..${path.sep}`)) {
-    throw new Error(`Unsafe R2 key: ${key}`);
-  }
-  if (path.isAbsolute(relative)) {
-    throw new Error(`Unsafe R2 key: ${key}`);
-  }
-
-  return destination;
-}
-
-export async function verifyLocalFile(
-  filePath: string,
-  candidate: R2ObjectMetadata,
-): Promise<LocalVerification> {
-  let fileStats: Awaited<ReturnType<typeof stat>>;
-  try {
-    fileStats = await stat(filePath);
-  } catch (error) {
-    if (isNodeError(error, 'ENOENT')) {
-      return { status: 'missing' };
-    }
-    throw error;
-  }
-
-  if (!fileStats.isFile()) {
-    return { reason: 'Local path is not a regular file', status: 'mismatch' };
-  }
-  if (fileStats.size !== candidate.size) {
-    return {
-      reason: `Local size ${fileStats.size} does not match R2 size ${candidate.size}`,
-      status: 'mismatch',
-    };
-  }
-
-  const expectedMd5 = simpleEtagMd5(candidate.etag);
-  if (!expectedMd5) {
-    return {
-      reason: 'R2 ETag is not a simple MD5 checksum',
-      status: 'unverifiable',
-    };
-  }
-
-  const actualMd5 = await md5File(filePath);
-  if (actualMd5 !== expectedMd5) {
-    return {
-      reason: `Local MD5 ${actualMd5} does not match R2 ETag ${expectedMd5}`,
-      status: 'mismatch',
-    };
-  }
-
-  return { status: 'verified' };
-}
-
-export async function downloadWithoutOverwrite(
-  client: CleanupR2Client,
-  candidate: R2ObjectMetadata,
-  downloadDir: string,
-): Promise<DownloadResult> {
-  const destination = resolveLocalObjectPath(
-    downloadDir,
-    candidate.bucket,
-    candidate.key,
-  );
-  let temporaryPath: string | undefined;
-
-  try {
-    const existing = await verifyLocalFile(destination, candidate);
-    if (existing.status !== 'missing') {
-      let status: DownloadResult['status'] = 'failed';
-      if (existing.status === 'verified') {
-        status = 'exists';
-      } else if (existing.status === 'unverifiable') {
-        status = 'unverifiable';
-      }
-      return {
-        destination,
-        reason: existing.reason,
-        status,
-      };
-    }
-
-    await mkdir(path.dirname(destination), { recursive: true });
-    temporaryPath = `${destination}.partial-${randomUUID()}`;
-    const body = await client.getObject(
-      candidate.bucket,
-      candidate.key,
-      candidate.etag,
-    );
-    await pipeline(
-      Readable.from(body),
-      createByteLimitTransform(candidate.size),
-      createWriteStream(temporaryPath, { flags: 'wx' }),
-    );
-
-    const verification = await verifyLocalFile(temporaryPath, candidate);
-    if (
-      verification.status === 'missing' ||
-      verification.status === 'mismatch'
-    ) {
-      return {
-        destination,
-        reason: verification.reason,
-        status: 'failed',
-      };
-    }
-
-    await finalizeWithoutOverwrite(temporaryPath, destination);
-    if (verification.status === 'unverifiable') {
-      return {
-        destination,
-        reason: verification.reason,
-        status: 'unverifiable',
-      };
-    }
-    return { destination, status: 'downloaded' };
-  } catch (error) {
-    return {
-      destination,
-      reason: error instanceof Error ? error.message : String(error),
-      status: 'failed',
-    };
-  } finally {
-    if (temporaryPath) {
-      await rm(temporaryPath, { force: true }).catch(() => undefined);
-    }
-  }
 }
 
 export async function deleteCandidatesInBatches(
@@ -1200,18 +854,6 @@ function compareLocationSummary(
   );
 }
 
-function compareOldestFirst(
-  left: R2ObjectMetadata,
-  right: R2ObjectMetadata,
-): number {
-  return (
-    new Date(left.lastModified).getTime() -
-      new Date(right.lastModified).getTime() ||
-    left.bucket.localeCompare(right.bucket) ||
-    left.key.localeCompare(right.key)
-  );
-}
-
 function assertBucketName(bucket: string): void {
   if (
     !bucket ||
@@ -1236,55 +878,4 @@ function assertIsoDate(value: unknown, label: string): asserts value is string {
 
 function isRecord(value: unknown): value is Record<string, unknown> {
   return typeof value === 'object' && value !== null && !Array.isArray(value);
-}
-
-function isNodeError(
-  error: unknown,
-  code: string,
-): error is NodeJS.ErrnoException {
-  return error instanceof Error && 'code' in error && error.code === code;
-}
-
-function createByteLimitTransform(maxBytes: number): Transform {
-  let bytes = 0;
-  return new Transform({
-    transform(chunk: Buffer, _encoding, callback) {
-      bytes += chunk.byteLength;
-      if (bytes > maxBytes) {
-        callback(
-          new Error(
-            `R2 response exceeded the expected size of ${maxBytes} bytes`,
-          ),
-        );
-        return;
-      }
-      callback(null, chunk);
-    },
-  });
-}
-
-async function finalizeWithoutOverwrite(
-  temporaryPath: string,
-  destination: string,
-): Promise<void> {
-  try {
-    await link(temporaryPath, destination);
-  } catch (error) {
-    const hardLinksUnsupported =
-      isNodeError(error, 'EPERM') ||
-      isNodeError(error, 'ENOTSUP') ||
-      isNodeError(error, 'EOPNOTSUPP');
-    if (!hardLinksUnsupported) {
-      throw error;
-    }
-    await copyFile(temporaryPath, destination, constants.COPYFILE_EXCL);
-  }
-}
-
-async function md5File(filePath: string): Promise<string> {
-  const hash = createHash('md5');
-  for await (const chunk of createReadStream(filePath)) {
-    hash.update(chunk);
-  }
-  return hash.digest('hex');
 }
